Reorganised the source code of the typing phase
