+++ src/main.rs
@@ -1,5 +1,5 @@
 use parser::FileObj;
-use typing::{flatten_const, type_ast};
+use typing::{flatten_const, simplify_ast, type_ast};
 
 pub mod parser;
 mod typing;
@@ -8,9 +8,11 @@ mod compiler;
 
 fn main() {
     match FileObj::new(
-        "/Users/emmanuel/Documents/developement/shared-projects/reactive-sys/CoqoZelus/tests/test_type",
+        "/Users/emmanuel/Documents/developement/shared-projects/reactive-sys/CoqoZelus/tests/test",
     ) {
         Ok(obj) => {
+            println!("{:?}", obj.get_ast());
+            let obj = simplify_ast(obj).unwrap();
             println!("{:?}", obj.get_ast());
             let obj =  type_ast(obj).unwrap();
             println!("{:?}", obj.get_ast());
