+++ tests/test
@@ -8,9 +8,8 @@ const
 	g = 10;
 	h = 34;
 
-
 node MonJolieNoeudQueJaime (x, y: int) returns () ;
 let
 	x = 0;
-	y = g+h;
+	y = g -> y;
 tel
